Add Extended Excel Processor

+++ src/main/java/com/z20let/mitigia/repository/EnergyConsumptionRepository.java
@@ -3,6 +3,6 @@ package com.z20let.mitigia.repository;
 import com.z20let.mitigia.model.EnergyConsumption;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnergyConsumptionRepository extends JpaRepository<EnergyConsumption, Long> {
+public interface EnergyConsumptionRepository extends JpaRepository<EnergyConsumption, Integer> {
 
 }
